ui: add shared table row click guard and reuse in InfiniteVirtualTable
